Handle all MSVC warnings
Summary: related to T13409

Reviewers: ivica

Reviewed By: ivica

Subscribers: miljen, iljazovic

Differential Revision: https://repo.mireo.local/D27152

+++ include/async_mqtt5/detail/control_packet.hpp
@@ -125,7 +125,7 @@ class packet_id_allocator {
 
 public:
 	packet_id_allocator() {
-		_free_ids.emplace_back(MAX_PACKET_ID, 0);
+		_free_ids.emplace_back(MAX_PACKET_ID, uint16_t(0));
 	}
 
 	uint16_t allocate() {
